Refactor: Move duration parsing to utils and fix bugs
