+++ src/components/upstreams/UpstreamDetailsDrawer.tsx
@@ -12,30 +12,7 @@ import { formatDistanceToNow } from "date-fns";
 import { useNavigate } from "react-router-dom";
 import { useToast } from "@/hooks/use-toast";
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis } from "recharts";
-
-/**
- * Parse a Caddy duration string to milliseconds
- * Supports: ms, s, m, h
- * Examples: "2000ms" -> 2000, "5s" -> 5000, "2m" -> 120000
- */
-function parseDurationToMs(duration: string): number {
-  const match = duration.match(/^(\d+(?:\.\d+)?)(ms|s|m|h)$/);
-  if (!match) {
-    console.warn(`Unable to parse duration: ${duration}`);
-    return parseInt(duration) || 0;
-  }
-  
-  const value = parseFloat(match[1]);
-  const unit = match[2];
-  
-  switch (unit) {
-    case 'ms': return value;
-    case 's': return value * 1000;
-    case 'm': return value * 60000;
-    case 'h': return value * 3600000;
-    default: return value;
-  }
-}
+import { parseDurationToMs } from "@/lib/utils";
 
 export function UpstreamDetailsDrawer({ upstream, instanceId, open, onClose, onTestHealth }: UpstreamDetailsDrawerProps) {
   const navigate = useNavigate();
